🚚  Move getLicense and getEslintConfig to /utils/
