+++ src/utils/getLicense.ts
